# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Enpoints para uso externo. API Entidad de credito B realizada por el Grupo 4, Integracion de aplicaciones | Segundo Cuatrimestre 2021",
    "version": "1.0.0",
    "title": "Swagger Entidad de credito B",
    "license": {
      "name": "UADE",
      "url": "https://www.uade.edu.ar/"
    }
  },
  "host": "ia-grupo4-backend.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Cobro",
      "description": "Endpoint para generacion de cobros"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/users/agregarMovimiento": {
      "post": {
        "tags": [
          "Cobro"
        ],
        "summary": "Agrega un nuevo cobro",
        "description": "",
        "operationId": "agregarMovimiento",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Los cobros deben ser registrados como un movimiento",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cobro"
            }
          }
        ],
        "responses": {
          "201": {
            "data": "String con el ID del movimiento generado",
            "status": 201,
            "message": "El pago se proceso correctamente"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "Negocio inexistente"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "Cliente inexistente"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "El monto debe ser mayor a 0"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "Número de tarjeta incorrecto"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "Codigo de seguridad incorrecto"
          },
          "400": {
            "data": "null",
            "status": 400,
            "message": "El monto supera el límite disponible"
          }
        }
      }
    }
  },
  "definitions": {
    "Cobro": {
      "type": "object",
      "properties": {
        "dnicuilUsuario": {
          "type": "string",
          "description": "Se puede enviar tanto el DNI como el CUIL del usuario. "
        },
        "cuitNegocio": {
          "type": "string",
          "description": "Minimo 11 caracteres valida contra los comercios registrados"
        },
        "numerotarjeta": {
          "type": "string",
          "description": "16 caracteres valida contra las tarjetas de los clientes"
        },
        "monto": {
          "type": "integer",
          "format": "int32",
          "description": "El monto no debe ser superior al asignado"
        },
        "codigoseguridad": {
          "type": "integer",
          "format": "int32",
          "description": "El codigo de seguridad de esa tarjeta en particular"
        }
      }
    }
  }
}